+++ examples/nextjs/app/shaka-video/page.tsx
@@ -1,5 +1,6 @@
 import { Metadata } from 'next';
 import ShakaVideo from 'shaka-video-element/react';
+import Player from '../player';
 
 export const metadata: Metadata = {
   title: 'Shaka Video - Media Elements',
@@ -9,14 +10,11 @@ export default function Page() {
   return (
     <>
       <section>
-        <ShakaVideo
-          className="video"
+        <Player
+          as={ShakaVideo}
           src="https://stream.mux.com/Sc89iWAyNkhJ3P1rQ02nrEdCFTnfT01CZ2KmaEcxXfB008.m3u8"
           poster="https://image.mux.com/Sc89iWAyNkhJ3P1rQ02nrEdCFTnfT01CZ2KmaEcxXfB008/thumbnail.webp?time=13"
-          controls
-          crossOrigin=""
-          playsInline
-        ></ShakaVideo>
+        />
       </section>
     </>
   );