+++ examples/nextjs/app/sidebar-nav.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { usePathname } from 'next/navigation'
-import Link from 'next/link';
+import Link from './link';
 
 export default function SidebarNav() {
   const pathname = usePathname()
@@ -40,6 +40,9 @@ export default function SidebarNav() {
       </li>
       <li>
         <Link className={`link ${pathname === '/videojs-video' ? 'active' : ''}`} href="/videojs-video">videojs-video</Link>
+      </li>
+      <li>
+        <Link className={`link ${pathname === '/twitch-video' ? 'active' : ''}`} href="/twitch-video">twitch-video</Link>
       </li>
       <li>
         <Link className={`link ${pathname === '/spotify-audio' ? 'active' : ''}`} href="/spotify-audio">spotify-audio</Link>