+++ packages/spotify-audio-element/CHANGELOG.md
@@ -5,6 +5,42 @@
 [1]: https://www.npmjs.com/package/spotify-audio-element?activeTab=versions
 
 
+## [1.0.2](https://github.com/muxinc/media-elements/compare/spotify-audio-element@1.0.1...spotify-audio-element@1.0.2) (2025-07-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * build-react-wrapper bumped from ^0.2.1 to ^0.2.2
+
+## [1.0.1](https://github.com/muxinc/media-elements/compare/spotify-audio-element@1.0.0...spotify-audio-element@1.0.1) (2025-07-11)
+
+
+### Bug Fixes
+
+* update iframe src with config bug ([3fa1fe8](https://github.com/muxinc/media-elements/commit/3fa1fe82928d077ab3cb1b060377fc8866455b31))
+
+## [1.0.0](https://github.com/muxinc/media-elements/compare/spotify-audio-element@0.1.14...spotify-audio-element@1.0.0) (2025-07-10)
+
+
+### ⚠ BREAKING CHANGES
+
+* remove attributes in favor of config property
+
+### Features
+
+* use config instead of attributes ([#139](https://github.com/muxinc/media-elements/issues/139)) ([f4c8610](https://github.com/muxinc/media-elements/commit/f4c86102e1e07f3eb2d63fa5370bfb1b9061dc5d))
+
+## [0.1.14](https://github.com/muxinc/media-elements/compare/spotify-audio-element@0.1.13...spotify-audio-element@0.1.14) (2025-06-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * build-react-wrapper bumped from ^0.2.0 to ^0.2.1
+
 ## [0.1.13](https://github.com/muxinc/media-elements/compare/spotify-audio-element@0.1.12...spotify-audio-element@0.1.13) (2025-03-12)