+++ packages/youtube-video-element/CHANGELOG.md
@@ -5,6 +5,38 @@
 [1]: https://www.npmjs.com/package/youtube-video-element?activeTab=versions
 
 
+## [1.6.1](https://github.com/muxinc/media-elements/compare/youtube-video-element@1.6.0...youtube-video-element@1.6.1) (2025-07-14)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * build-react-wrapper bumped from ^0.2.1 to ^0.2.2
+
+## [1.6.0](https://github.com/muxinc/media-elements/compare/youtube-video-element@1.5.3...youtube-video-element@1.6.0) (2025-07-02)
+
+
+### Features
+
+* **youtube-video:** add playlist support ([#134](https://github.com/muxinc/media-elements/issues/134)) ([34f7469](https://github.com/muxinc/media-elements/commit/34f7469b4b8204d92346ccb911c6a734632621d3))
+
+## [1.5.3](https://github.com/muxinc/media-elements/compare/youtube-video-element@1.5.2...youtube-video-element@1.5.3) (2025-06-26)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * build-react-wrapper bumped from ^0.2.0 to ^0.2.1
+
+## [1.5.2](https://github.com/muxinc/media-elements/compare/youtube-video-element@1.5.1...youtube-video-element@1.5.2) (2025-06-26)
+
+
+### Bug Fixes
+
+* **youtube-video-element:** support -nocookie ([#125](https://github.com/muxinc/media-elements/issues/125)) ([159199d](https://github.com/muxinc/media-elements/commit/159199d819273b201c7ce370801724b4167d7949))
+
 ## [1.5.1](https://github.com/muxinc/media-elements/compare/youtube-video-element@1.5.0...youtube-video-element@1.5.1) (2025-03-14)